Fixed LPF and Bitcrush to run on STM

+++ src/processors/lpf.h
@@ -12,7 +12,7 @@ public:
 	{
 		float output = 0;
 		float fSlow0 = (1.0f / q);
-		float fSlow1 = tan((fConst0 * cutoff));
+		float fSlow1 = tanf((fConst0 * cutoff));
 		float fSlow2 = (1.0f / fSlow1);
 		float fSlow3 = (1.0f / (((fSlow0 + fSlow2) / fSlow1) + 1.0f));
 		float fSlow4 = (((fSlow2 - fSlow0) / fSlow1) + 1.0f);
